+++ app/Main.hs
@@ -12,8 +12,8 @@ import Epidemic.BirthDeathSamplingOccurrence
 
 main :: IO ()
 main = do
-  events <- birthDeathSamplingOccurrenceSimulation $ birthDeathSamplingOccurrenceConfig 6 1.3 0.1 0.1 0.2
+  events <- birthDeathSamplingOccurrenceSimulation $ birthDeathSamplingOccurrenceConfig 6 (1.3,0.1,0.1,0.2)
   L.writeFile "demo-output.csv" (encode events)
   B.writeFile "demo-output.json" (eventsAsJsonTree events)
-  print $ show . sampleTreeEvents . sampleTree $ transmissionTree (reverse events) (Person 1)
+  L.writeFile "demo-output-sample-tree.csv" $ encode . sampleTreeEvents . sampleTree $ transmissionTree events (Person 1)
   return ()